\name{remove.oldtmps}
\alias{remove.oldtmps}
\title{Easy removal of old, obsolete tempfolders from previous sessions.}
\usage{
  remove.oldtmps()
}
\arguments{
  \item{recursive}{logical. Should directories be deleted recursively?}
  \item{force}{logical. Should permissions be changed (if possible) to allow the file or directory to be removed?}
  \item{expand}{logical. Should wildcards and tilde be expanded?}
}
\description{
  Easy removal of old, obsolete tempfolders from previous sessions.
}
\examples{
  library(pizzR)

  # not executed for safety
  #remove.oldtmps()
}
